+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+- Add support for Rich Text using `AutoSizeText.rich()` with one or multiple `TextSpan`s.
+- Improved text size calculation (using `textScaleFactor`)
+
 ## 0.1.0
 - Fixed documentation (thanks @g-balas)
 - Added tests